Organization and Settings file
- Divided code into helpers and main
- Created settings file on setupDirectory
